add basic tests for component updates with `useState` and `useEffect`
